+++ packages/ListBox/src/components/Options/Options.js
@@ -5,6 +5,8 @@ export default function Options(props) {
   const { children, isPopoverOpen } = props;
   let index = -1;
   return React.Children.map(children, child => {
+    if (child === null) return null;
+
     const { displayName = null } = child.type;
 
     if (child.type && isWhiteListed(displayName)) {