+++ packages/Breadcrumbs/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed the potential mapping children error.
+
 ## [0.0.2-alpha.0] - 2020-10-23
 
 ### Added